Refactor modules structure, add GradCAM to object detection module
