+++ example/gradcam/main.py
@@ -10,11 +10,21 @@ import torch
 import torchvision
 from deep_utils import Box
 from PIL import Image
-from src.gradcam import GradCAMObjectDetection, ObjectDetectionOutput
-from src.model import WrapperYOLOv5ObjectDetectionModel
-from src.object_detector import SSDObjectDetector, YOLOv5ObjectDetector, get_yolo_layer
 from torchvision.models._meta import _COCO_CATEGORIES
 from torchvision.models.detection import SSD300_VGG16_Weights
+
+from foxai.explainer.computer_vision.object_detection.gradcam import (
+    GradCAMObjectDetection,
+    ObjectDetectionOutput,
+)
+from foxai.explainer.computer_vision.object_detection.model import (
+    WrapperYOLOv5ObjectDetectionModel,
+)
+from foxai.explainer.computer_vision.object_detection.object_detector import (
+    SSDObjectDetector,
+    YOLOv5ObjectDetector,
+    get_yolo_layer,
+)
 
 
 def draw_heatmap_in_bbox(
