+++ foxai/explainer/__init__.py
@@ -1,29 +1,48 @@
 # pylint: disable = missing-module-docstring
-from foxai.explainer.conductance import LayerConductanceCVExplainer
-from foxai.explainer.deconv import DeconvolutionCVExplainer
-from foxai.explainer.deeplift import DeepLIFTCVExplainer, LayerDeepLIFTCVExplainer
-from foxai.explainer.deeplift_shap import (
+from foxai.explainer.computer_vision.image_classification.conductance import (
+    LayerConductanceCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.deconv import (
+    DeconvolutionCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.deeplift import (
+    DeepLIFTCVExplainer,
+    LayerDeepLIFTCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.deeplift_shap import (
     DeepLIFTSHAPCVExplainer,
     LayerDeepLIFTSHAPCVExplainer,
 )
-from foxai.explainer.gradcam import GuidedGradCAMCVExplainer, LayerGradCAMCVExplainer
-from foxai.explainer.gradient_shap import (
+from foxai.explainer.computer_vision.image_classification.gradcam import (
+    GuidedGradCAMCVExplainer,
+    LayerGradCAMCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.gradient_shap import (
     GradientSHAPCVExplainer,
     LayerGradientSHAPCVExplainer,
 )
-from foxai.explainer.guided_backprop import GuidedBackpropCVExplainer
-from foxai.explainer.input_x_gradient import (
+from foxai.explainer.computer_vision.image_classification.guided_backprop import (
+    GuidedBackpropCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.input_x_gradient import (
     InputXGradientCVExplainer,
     LayerInputXGradientCVExplainer,
 )
-from foxai.explainer.integrated_gradients import (
+from foxai.explainer.computer_vision.image_classification.integrated_gradients import (
     IntegratedGradientsCVExplainer,
     LayerIntegratedGradientsCVExplainer,
 )
-from foxai.explainer.lrp import LayerLRPCVExplainer, LRPCVExplainer
-from foxai.explainer.noise_tunnel import (
+from foxai.explainer.computer_vision.image_classification.lrp import (
+    LayerLRPCVExplainer,
+    LRPCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.noise_tunnel import (
     LayerNoiseTunnelCVExplainer,
     NoiseTunnelCVExplainer,
 )
-from foxai.explainer.occlusion import OcclusionCVExplainer
-from foxai.explainer.saliency import SaliencyCVExplainer
+from foxai.explainer.computer_vision.image_classification.occlusion import (
+    OcclusionCVExplainer,
+)
+from foxai.explainer.computer_vision.image_classification.saliency import (
+    SaliencyCVExplainer,
+)
