+++ foxai/explainer/computer_vision/image_classification/integrated_gradients.py
@@ -1,23 +1,25 @@
-"""File with Gradient SHAP algorithm explainer classes.
+"""File with Integrated Gradients algorithm explainer classes.
 
-Based on https://github.com/pytorch/captum/blob/master/captum/attr/_core/gradient_shap.py
-and https://github.com/pytorch/captum/blob/master/captum/attr/_core/layer/layer_gradient_shap.py.
+Based on https://github.com/pytorch/captum/blob/master/captum/attr/_core/integrated_gradients.py
+and https://github.com/pytorch/captum/blob/master/captum/attr/_core/layer/layer_integrated_gradients.py.
 """
 
 from abc import abstractmethod
-from typing import Any, Optional, Tuple, Union
+from typing import Any, Optional, Union
 
 import torch
 from captum._utils.typing import TargetType
-from captum.attr import GradientShap, LayerGradientShap
+from captum.attr import IntegratedGradients, LayerIntegratedGradients
 
 from foxai.array_utils import validate_result
-from foxai.explainer.base_explainer import CVExplainer
-from foxai.explainer.model_utils import get_last_conv_model_layer
+from foxai.explainer.computer_vision.image_classification.base_explainer import (
+    Explainer,
+)
+from foxai.explainer.computer_vision.model_utils import get_last_conv_model_layer
 
 
-class BaseGradientSHAPCVExplainer(CVExplainer):
-    """Base Gradient SHAP algorithm explainer."""
+class BaseIntegratedGradientsCVExplainer(Explainer):
+    """Base Integrated Gradients algorithm explainer."""
 
     @abstractmethod
     def create_explainer(
@@ -25,7 +27,7 @@ class BaseGradientSHAPCVExplainer(CVExplainer):
         model: torch.nn.Module,
         multiply_by_inputs: bool = True,
         **kwargs,
-    ) -> Union[GradientShap, LayerGradientShap]:
+    ) -> Union[IntegratedGradients, LayerIntegratedGradients]:
         """Create explainer object.
 
         Args:
@@ -34,16 +36,20 @@ class BaseGradientSHAPCVExplainer(CVExplainer):
             multiply_by_inputs: Indicates whether to factor
                 model inputs' multiplier in the final attribution scores.
                 In the literature this is also known as local vs global
-                attribution. If inputs' multiplier isn't factored in
-                then this type of attribution method is also called local
+                attribution. If inputs' multiplier isn't factored in,
+                then that type of attribution method is also called local
                 attribution. If it is, then that type of attribution
                 method is called global.
                 More detailed can be found here:
                 https://arxiv.org/abs/1711.06104
 
-                In case of gradient shap, if `multiply_by_inputs`
-                is set to True, the sensitivity scores of scaled inputs
-                are being multiplied by (inputs - baselines).
+                In case of integrated gradients, if `multiply_by_inputs`
+                is set to True, final sensitivity scores are being multiplied by
+                (inputs - baselines).
+
+                In case of layer integrated gradients, if `multiply_by_inputs`
+                is set to True, final sensitivity scores are being multiplied by
+                layer activations for inputs - layer activations for baselines.
 
         Returns:
             Explainer object.
@@ -55,19 +61,20 @@ class BaseGradientSHAPCVExplainer(CVExplainer):
         input_data: torch.Tensor,
         pred_label_idx: TargetType = None,
         baselines: Union[None, int, float, torch.Tensor] = None,
-        n_samples: int = 5,
-        stdevs: Union[float, Tuple[float, ...]] = 0.0,
         additional_forward_args: Any = None,
+        n_steps: int = 50,
+        method: str = "gausslegendre",
+        internal_batch_size: Union[None, int] = None,
         attribute_to_layer_input: bool = False,
         **kwargs,
     ) -> torch.Tensor:
-        """Generate model's attributes with Gradient SHAP algorithm explainer.
+        """Generate model's attributes with Integrated Gradients algorithm explainer.
 
         Args:
             model: The forward function of the model or any
                 modification of it.
-            input_data: Input for which SHAP attribution
-                values are computed. If `forward_func` takes a single
+            input_data: Input for which layer integrated
+                gradients are computed. If forward_func takes a single
                 tensor as input, a single input tensor should be provided.
             pred_label_idx: Output indices for
                 which gradients are computed (for classification cases,
@@ -95,44 +102,58 @@ class BaseGradientSHAPCVExplainer(CVExplainer):
 
                 Default: None
             baselines:
-                Baselines define the starting point from which expectation
+                Baselines define the starting point from which integral
                 is computed and can be provided as:
 
                 - a single tensor, if inputs is a single tensor, with
-                    the first dimension equal to the number of examples
-                    in the baselines' distribution. The remaining dimensions
-                    must match with input tensor's dimension starting from
-                    the second dimension.
+                    exactly the same dimensions as inputs or the first
+                    dimension is one and the remaining dimensions match
+                    with inputs.
 
-                It is recommended that the number of samples in the baselines'
-                tensors is larger than one.
-            n_samples: The number of randomly generated examples
-                per sample in the input batch. Random examples are
-                generated by adding gaussian random noise to each sample.
-                Default: `5` if `n_samples` is not provided.
-            stdevs: The standard deviation
-                of gaussian noise with zero mean that is added to each
-                input in the batch. If `stdevs` is a single float value
-                then that same value is used for all inputs. If it is
-                a tuple, then it must have the same length as the inputs
-                tuple. In this case, each stdev value in the stdevs tuple
-                corresponds to the input with the same index in the inputs
-                tuple.
-                Default: 0.0
+                - a single scalar, if inputs is a single tensor, which will
+                    be broadcasted for each input value in input tensor.
+
+                In the cases when `baselines` is not provided, we internally
+                use zero scalar corresponding to each input tensor.
+                Default: None
             additional_forward_args: If the forward function
                 requires additional arguments other than the inputs for
                 which attributions should not be computed, this argument
-                can be provided. It can contain a tuple of ND tensors or
-                any arbitrary python type of any shape.
-                In case of the ND tensor the first dimension of the
-                tensor must correspond to the batch size. It will be
-                repeated for each `n_steps` for each randomly generated
-                input sample.
+                can be provided. It must be either a single additional
+                argument of a Tensor or arbitrary (non-tuple) type or a
+                tuple containing multiple additional arguments including
+                tensors or any arbitrary python types. These arguments
+                are provided to forward_func in order following the
+                arguments in inputs.
 
-                Note that the gradients are not computed with respect
+                For a tensor, the first dimension of the tensor must
+                correspond to the number of examples. It will be
+                repeated for each of `n_steps` along the integrated
+                path. For all other types, the given argument is used
+                for all forward evaluations.
+
+                Note that attributions are not computed with respect
                 to these arguments.
                 Default: None
-            attribute_to_layer_input (bool, optional): Indicates whether to
+            n_steps: The number of steps used by the approximation
+                method. Default: 50.
+            method: Method for approximating the integral,
+                one of `riemann_right`, `riemann_left`, `riemann_middle`,
+                `riemann_trapezoid` or `gausslegendre`.
+                Default: `gausslegendre` if no method is provided.
+            internal_batch_size: Divides total #steps * #examples
+                data points into chunks of size at most internal_batch_size,
+                which are computed (forward / backward passes)
+                sequentially. internal_batch_size must be at least equal to
+                #examples.
+
+                For DataParallel models, each batch is split among the
+                available devices, so evaluations on each available
+                device contain internal_batch_size / num_devices examples.
+                If internal_batch_size is None, then all evaluations are
+                processed in one batch.
+                Default: None
+            attribute_to_layer_input: Indicates whether to
                 compute the attribution with respect to the layer input
                 or output. If `attribute_to_layer_input` is set to True
                 then the attributions will be computed with respect to
@@ -146,65 +167,75 @@ class BaseGradientSHAPCVExplainer(CVExplainer):
                 Default: False
 
         Returns:
-            Attribution score computed based on GradientSHAP with respect
-            to each input feature. Attributions will always be
-            the same size as the provided inputs, with each value
-            providing the attribution of the corresponding input index.
-            If a single tensor is provided as inputs, a single tensor is
-            returned. If a tuple is provided for inputs, a tuple of
-            corresponding sized tensors is returned.
+            Integrated gradients with respect to `layer`'s inputs
+            or outputs. Attributions will always be the same size and
+            dimensionality as the input or output of the given layer,
+            depending on whether we attribute to the inputs or outputs
+            of the layer which is decided by the input flag
+            `attribute_to_layer_input`.
+
+            For a single layer, attributions are returned in a tuple if
+            the layer inputs / outputs contain multiple tensors,
+            otherwise a single tensor is returned.
+
+            For multiple layers, attributions will always be
+            returned as a list. Each element in this list will be
+            equivalent to that of a single layer output, i.e. in the
+            case that one layer, in the given layers, inputs / outputs
+            multiple tensors: the corresponding output element will be
+            a tuple of tensors. The ordering of the outputs will be
+            the same order as the layers given in the constructor.
 
         Raises:
             RuntimeError: if attribution has shape (0).
         """
         layer: Optional[torch.nn.Module] = kwargs.get("layer", None)
-        gradient_shap = self.create_explainer(model=model, layer=layer)
+        integrated_gradients = self.create_explainer(model=model, layer=layer)
 
         # defining baseline distribution of images
         if baselines is None:
             baselines = torch.randn(
-                (
-                    2 * input_data.shape[0],
-                    *input_data.shape[1:],
-                ),
+                input_data.shape,
                 requires_grad=True,
                 device=input_data.device,
             )
 
-        if isinstance(gradient_shap, LayerGradientShap):
-            attributions = gradient_shap.attribute(
+        if isinstance(integrated_gradients, LayerIntegratedGradients):
+            attributions = integrated_gradients.attribute(
                 input_data,
-                n_samples=n_samples,
-                stdevs=stdevs,
-                baselines=baselines,
                 target=pred_label_idx,
+                n_steps=n_steps,
+                baselines=baselines,
                 return_convergence_delta=False,
                 additional_forward_args=additional_forward_args,
+                method=method,
+                internal_batch_size=internal_batch_size,
                 attribute_to_layer_input=attribute_to_layer_input,
             )
         else:
-            attributions = gradient_shap.attribute(
+            attributions = integrated_gradients.attribute(
                 input_data,
-                n_samples=n_samples,
-                stdevs=stdevs,
-                baselines=baselines,
                 target=pred_label_idx,
+                baselines=baselines,
+                n_steps=n_steps,
                 return_convergence_delta=False,
                 additional_forward_args=additional_forward_args,
+                method=method,
+                internal_batch_size=internal_batch_size,
             )
         validate_result(attributions=attributions)
         return attributions
 
 
-class GradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
-    """Gradient SHAP algorithm explainer."""
+class IntegratedGradientsCVExplainer(BaseIntegratedGradientsCVExplainer):
+    """Integrated Gradients algorithm explainer."""
 
     def create_explainer(
         self,
         model: torch.nn.Module,
         multiply_by_inputs: bool = True,
         **kwargs,
-    ) -> Union[GradientShap, LayerGradientShap]:
+    ) -> Union[IntegratedGradients, LayerIntegratedGradients]:
         """Create explainer object.
 
         Args:
@@ -213,29 +244,29 @@ class GradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
             multiply_by_inputs: Indicates whether to factor
                 model inputs' multiplier in the final attribution scores.
                 In the literature this is also known as local vs global
-                attribution. If inputs' multiplier isn't factored in
-                then this type of attribution method is also called local
+                attribution. If inputs' multiplier isn't factored in,
+                then that type of attribution method is also called local
                 attribution. If it is, then that type of attribution
                 method is called global.
                 More detailed can be found here:
                 https://arxiv.org/abs/1711.06104
 
-                In case of gradient shap, if `multiply_by_inputs`
-                is set to True, the sensitivity scores of scaled inputs
-                are being multiplied by (inputs - baselines).
+                In case of integrated gradients, if `multiply_by_inputs`
+                is set to True, final sensitivity scores are being multiplied by
+                (inputs - baselines).
 
         Returns:
             Explainer object.
         """
 
-        return GradientShap(
+        return IntegratedGradients(
             forward_func=model,
             multiply_by_inputs=multiply_by_inputs,
         )
 
 
-class LayerGradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
-    """Layer Gradient SHAP algorithm explainer."""
+class LayerIntegratedGradientsCVExplainer(BaseIntegratedGradientsCVExplainer):
+    """Layer Integrated Gradients algorithm explainer."""
 
     def create_explainer(
         self,
@@ -243,7 +274,7 @@ class LayerGradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
         multiply_by_inputs: bool = True,
         layer: Optional[torch.nn.Module] = None,
         **kwargs,
-    ) -> Union[GradientShap, LayerGradientShap]:
+    ) -> Union[IntegratedGradients, LayerIntegratedGradients]:
         """Create explainer object.
 
         Uses parameter `layer` from `kwargs`. If not provided function will call
@@ -263,16 +294,15 @@ class LayerGradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
             multiply_by_inputs: Indicates whether to factor
                 model inputs' multiplier in the final attribution scores.
                 In the literature this is also known as local vs global
-                attribution. If inputs' multiplier isn't factored in
-                then this type of attribution method is also called local
+                attribution. If inputs' multiplier isn't factored in,
+                then that type of attribution method is also called local
                 attribution. If it is, then that type of attribution
                 method is called global.
                 More detailed can be found here:
                 https://arxiv.org/abs/1711.06104
 
-                In case of layer gradient shap, if `multiply_by_inputs`
-                is set to True, the sensitivity scores for scaled inputs
-                are being multiplied by
+                In case of layer integrated gradients, if `multiply_by_inputs`
+                is set to True, final sensitivity scores are being multiplied by
                 layer activations for inputs - layer activations for baselines.
 
         Returns:
@@ -281,12 +311,11 @@ class LayerGradientSHAPCVExplainer(BaseGradientSHAPCVExplainer):
         Raises:
             ValueError: if model does not contain conv layers.
         """
-
         if layer is None:
             layer = get_last_conv_model_layer(model=model)
 
-        return LayerGradientShap(
+        return LayerIntegratedGradients(
             forward_func=model,
-            layer=layer,
             multiply_by_inputs=multiply_by_inputs,
+            layer=layer,
         )
