+++ foxai/explainer/computer_vision/image_classification/noise_tunnel.py
@@ -11,11 +11,13 @@ from captum._utils.typing import TargetType
 from captum.attr import IntegratedGradients, LayerIntegratedGradients, NoiseTunnel
 
 from foxai.array_utils import validate_result
-from foxai.explainer.base_explainer import CVExplainer
-from foxai.explainer.model_utils import get_last_conv_model_layer
+from foxai.explainer.computer_vision.image_classification.base_explainer import (
+    Explainer,
+)
+from foxai.explainer.computer_vision.model_utils import get_last_conv_model_layer
 
 
-class BaseNoiseTunnelCVExplainer(CVExplainer):
+class BaseNoiseTunnelCVExplainer(Explainer):
     """Base Noise Tunnel algorithm explainer."""
 
     @abstractmethod
