+++ foxai/explainer/computer_vision/image_classification/saliency.py
@@ -1,57 +1,40 @@
-"""File with Guided Backpropagation algorithm explainer class.
+"""File with Saliency algorithm explainer classes.
 
-Based on https://github.com/pytorch/captum/blob/master/captum/attr/_core/guided_backprop_deconvnet.py.
+Based on https://github.com/pytorch/captum/blob/master/captum/attr/_core/saliency.py.
 """
 
-from abc import abstractmethod
 from typing import Any
 
 import torch
 from captum._utils.typing import TargetType
-from captum.attr import GuidedBackprop
+from captum.attr import Saliency
 
 from foxai.array_utils import validate_result
-from foxai.explainer.base_explainer import CVExplainer
-from foxai.explainer.model_utils import modify_modules
+from foxai.explainer.computer_vision.image_classification.base_explainer import (
+    Explainer,
+)
 
 
-class BaseGuidedBackpropCVExplainer(CVExplainer):
-    """Base Guided Backpropagation algorithm explainer."""
-
-    @abstractmethod
-    def create_explainer(
-        self,
-        model: torch.nn.Module,
-        **kwargs,
-    ) -> GuidedBackprop:
-        """Create explainer object.
-
-        Returns:
-            Explainer object.
-        """
+class SaliencyCVExplainer(Explainer):
+    """Saliency algorithm explainer."""
 
     def calculate_features(
         self,
         model: torch.nn.Module,
         input_data: torch.Tensor,
         pred_label_idx: TargetType = None,
+        abs_value: bool = True,
         additional_forward_args: Any = None,
-        **kwargs,
+        **kwargs,  # pylint: disable = (unused-argument)
     ) -> torch.Tensor:
-        """Generate model's attributes with Guided Backpropagation algorithm explainer.
+        """Generate model's attributes with Saliency algorithm explainer.
 
         Args:
             model: The forward function of the model or any
                 modification of it.
-            input_data: Input for which
-                attributions are computed. If forward_func takes a single
-                tensor as input, a single input tensor should be provided.
-                If forward_func takes multiple tensors as input, a tuple
-                of the input tensors should be provided. It is assumed
-                that for all given input tensors, dimension 0 corresponds
-                to the number of examples (aka batch size), and if
-                multiple input tensors are provided, the examples must
-                be aligned appropriately.
+            input_data: Input for which saliency
+                is computed. If forward_func takes a single tensor
+                as input, a single input tensor should be provided.
             pred_label_idx: Output indices for
                 which gradients are computed (for classification cases,
                 this is usually the target class).
@@ -77,22 +60,27 @@ class BaseGuidedBackpropCVExplainer(CVExplainer):
                     target for the corresponding example.
 
                 Default: None
+            abs_value: Returns absolute value of gradients if set
+                to True, otherwise returns the (signed) gradients if
+                False.
+                Default: True
             additional_forward_args: If the forward function
                 requires additional arguments other than the inputs for
                 which attributions should not be computed, this argument
                 can be provided. It must be either a single additional
-                argument of a Tensor or arbitrary (non-tuple) type or a tuple
-                containing multiple additional arguments including tensors
-                or any arbitrary python types. These arguments are provided to
-                forward_func in order, following the arguments in inputs.
+                argument of a Tensor or arbitrary (non-tuple) type or a
+                tuple containing multiple additional arguments including
+                tensors or any arbitrary python types. These arguments
+                are provided to forward_func in order following the
+                arguments in inputs.
                 Note that attributions are not computed with respect
                 to these arguments.
                 Default: None
 
         Returns:
-            The guided backprop gradients with respect to each
-            input feature. Attributions will always
-            be the same size as the provided inputs, with each value
+            The gradients with respect to each input feature.
+            Attributions will always be
+            the same size as the provided inputs, with each value
             providing the attribution of the corresponding input index.
             If a single tensor is provided as inputs, a single tensor is
             returned. If a tuple is provided for inputs, a tuple of
@@ -101,28 +89,13 @@ class BaseGuidedBackpropCVExplainer(CVExplainer):
         Raises:
             RuntimeError: if attribution has shape (0).
         """
-        guided_backprop = self.create_explainer(model=model)
-        attributions = guided_backprop.attribute(
+        saliency = Saliency(forward_func=model)
+
+        attributions = saliency.attribute(
             input_data,
             target=pred_label_idx,
+            abs=abs_value,
             additional_forward_args=additional_forward_args,
         )
         validate_result(attributions=attributions)
         return attributions
-
-
-class GuidedBackpropCVExplainer(BaseGuidedBackpropCVExplainer):
-    """Guided Backpropagation algorithm explainer."""
-
-    def create_explainer(
-        self,
-        model: torch.nn.Module,
-        **kwargs,
-    ) -> GuidedBackprop:
-        """Create explainer object.
-
-        Returns:
-            Explainer object.
-        """
-
-        return GuidedBackprop(model=modify_modules(model))
