+++ tests/test_explainers.py
@@ -9,8 +9,10 @@ import torch
 from torchvision import transforms
 
 from foxai import explainer
-from foxai.context_manager import FoXaiExplainer, Explainers, ExplainerWithParams
-from foxai.explainer.base_explainer import CVExplainerT
+from foxai.context_manager import Explainers, ExplainerWithParams, FoXaiExplainer
+from foxai.explainer.computer_vision.image_classification.base_explainer import (
+    CVExplainerT,
+)
 from foxai.logger import create_logger
 from tests.pickachu_image import pikachu_image
 from tests.sample_model import SampleModel
@@ -143,7 +145,9 @@ class TestExplainers:
             _, _ = xai_model(img_tensor)
 
 
-@patch("foxai.explainer.conductance.LayerConductance.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.conductance.LayerConductance.attribute"
+)
 def test_conductance_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -160,7 +164,9 @@ def test_conductance_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.deconv.Deconvolution.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.deconv.Deconvolution.attribute"
+)
 def test_deconvolution_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -176,7 +182,9 @@ def test_deconvolution_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.deeplift_shap.DeepLiftShap.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.deeplift_shap.DeepLiftShap.attribute"
+)
 def test_deepliftshap_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -192,7 +200,9 @@ def test_deepliftshap_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.deeplift_shap.LayerDeepLiftShap.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.deeplift_shap.LayerDeepLiftShap.attribute"
+)
 def test_layer_deepliftshap_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -208,7 +218,9 @@ def test_layer_deepliftshap_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.deeplift.DeepLift.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.deeplift.DeepLift.attribute"
+)
 def test_deeplift_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -224,7 +236,9 @@ def test_deeplift_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.deeplift.LayerDeepLift.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.deeplift.LayerDeepLift.attribute"
+)
 def test_layer_deeplift_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -240,7 +254,9 @@ def test_layer_deeplift_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.gradcam.GuidedGradCam.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.gradcam.GuidedGradCam.attribute"
+)
 def test_gradcam_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -257,7 +273,9 @@ def test_gradcam_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.gradcam.LayerGradCam.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.gradcam.LayerGradCam.attribute"
+)
 def test_layer_gradcam_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -274,7 +292,9 @@ def test_layer_gradcam_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.gradient_shap.GradientShap.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.gradient_shap.GradientShap.attribute"
+)
 def test_gradient_shap_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -290,7 +310,9 @@ def test_gradient_shap_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.gradient_shap.LayerGradientShap.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.gradient_shap.LayerGradientShap.attribute"
+)
 def test_layer_gradient_shap_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -306,7 +328,9 @@ def test_layer_gradient_shap_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.input_x_gradient.InputXGradient.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.input_x_gradient.InputXGradient.attribute"
+)
 def test_input_x_gradient_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -322,7 +346,9 @@ def test_input_x_gradient_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.input_x_gradient.LayerGradientXActivation.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.input_x_gradient.LayerGradientXActivation.attribute"
+)
 def test_layer_input_x_gradient_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -338,7 +364,7 @@ def test_layer_input_x_gradient_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.lrp.LRP.attribute")
+@patch("foxai.explainer.computer_vision.image_classification.lrp.LRP.attribute")
 def test_lrp_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -354,7 +380,7 @@ def test_lrp_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.lrp.LayerLRP.attribute")
+@patch("foxai.explainer.computer_vision.image_classification.lrp.LayerLRP.attribute")
 def test_layer_lrp_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -370,7 +396,9 @@ def test_layer_lrp_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.integrated_gradients.IntegratedGradients.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.integrated_gradients.IntegratedGradients.attribute"
+)
 def test_integrated_gradients_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -386,7 +414,9 @@ def test_integrated_gradients_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.integrated_gradients.LayerIntegratedGradients.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.integrated_gradients.LayerIntegratedGradients.attribute"
+)
 def test_layer_integrated_gradients_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -402,7 +432,9 @@ def test_layer_integrated_gradients_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.noise_tunnel.NoiseTunnel.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.noise_tunnel.NoiseTunnel.attribute"
+)
 def test_noise_tunnel_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -418,7 +450,9 @@ def test_noise_tunnel_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.noise_tunnel.NoiseTunnel.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.noise_tunnel.NoiseTunnel.attribute"
+)
 def test_layer_noise_tunnel_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -434,7 +468,9 @@ def test_layer_noise_tunnel_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.occlusion.Occlusion.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.occlusion.Occlusion.attribute"
+)
 def test_occulusion_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
@@ -450,7 +486,9 @@ def test_occulusion_raises_error_if_attributes_are_empty(
         )
 
 
-@patch("foxai.explainer.saliency.Saliency.attribute")
+@patch(
+    "foxai.explainer.computer_vision.image_classification.saliency.Saliency.attribute"
+)
 def test_saliency_raises_error_if_attributes_are_empty(
     fake_attribute: MagicMock,
 ) -> None:
